[Fix] Update ETL title & body
